pyproject.toml
[project]
name = "iway-dns-update"
version = "1.0.0"
description = "Update my DANE TLSA record"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pre-commit>=4.5.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruff>=0.14.13",
]

